tests/wpt/tests/css/css-grid/alignment/grid-align-baseline-table-001.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/7655#issuecomment-1248528627">
<style>
.target {
  display: grid;
  grid-auto-flow: column;
  justify-items: start;
  position: relative;
}
.target > div, td {
  background: hotpink;
  font-size: 20px;
  line-height: 0;
}
table { border-spacing: 0; }
td { padding: 5px; min-width: 5px; min-height: 5px; }
caption {
  margin: 10px;
  background: lime;
  block-size: 20px;
}
span {
  display: inline-block;
  width: 1em;
  height: 1em;
  outline: solid cyan 3px;
  outline-offset: -3px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.target > *')">

<div class="target" style="align-items: first baseline;">
  <div data-offset-y="45"><span></span></div>
  <table data-offset-y="0">
    <caption>top</caption>
    <tr style="vertical-align: baseline;">
      <td style="font-size: 10px;"><span></span><br><span></span></td>
      <td><span></span><br><span></span></td>
    </tr>
    <tr style="vertical-align: baseline;">
      <td><span></span><br><span></span></td>
      <td style="font-size: 10px;"><span></span><br><span></span></td>
    </tr>
    <caption style="caption-side: bottom;">bottom</caption>
  </table>
</div>

<div class="target" style="align-items: last baseline;">
  <div data-offset-y="95"><span></span></div>
  <table data-offset-y="0">
    <caption>top</caption>
    <tr style="vertical-align: baseline;">
      <td style="font-size: 10px;"><span></span><br><span></span></td>
      <td><span></span><br><span></span></td>
    </tr>
    <tr style="vertical-align: baseline;">
      <td><span></span><br><span></span></td>
      <td style="font-size: 10px;"><span></span><br><span></span></td>
    </tr>
    <caption style="caption-side: bottom;">bottom</caption>
  </table>
</div>

<div class="target" style="align-items: first baseline;">
  <div data-offset-y="20"><span></span></div>
  <table data-offset-y="0">
    <tr>
      <td style="border: solid 10px;"></td>
      <td style="border: solid 5px;"></td>
      <td style="border: solid 20px;"></td>
    </tr>
    <tr>
      <td style="border: solid 10px;"></td>
      <td style="border: solid 5px;"></td>
      <td style="border: solid 20px;"></td>
    </tr>
  </table>
</div>

<div class="target" style="align-items: last baseline;">
  <div data-offset-y="70"><span></span></div>
  <table data-offset-y="0">
    <tr>
      <td style="border: solid 10px;"></td>
      <td style="border: solid 5px;"></td>
      <td style="border: solid 20px;"></td>
    </tr>
    <tr>
      <td style="border: solid 10px;"></td>
      <td style="border: solid 5px;"></td>
      <td style="border: solid 20px;"></td>
    </tr>
  </table>
</div>
